wip
the current code shows a half-completed implementation of moving
generics out of ParselyWrite's type and into associated types.  Then
ParselyWrite can be used as a trait bound, which is useful when trying
to coerce types correctly when writing (e.g. the 'map' ambiguity
problem).  Doing this requires an additional type (WriteAdaptor) which
doesn't work in its current form: instead of a single impl defined in
parsely we'd need to generate impls dynamically for every type otherwise
we run into "can't-define-foreign-trait-on-foreign-type" problems when
we generate impls in a calling crate.