Better expression ergonomics (#5)
* change how StateSync trait is defined and make it a required supertrait of ParselyWrite

* allow context/sync expressions to return a raw value or a result

* refactor: rename sync_func -> sync_expr

* wip: add Assertion newtype to clean things up a bit

* wip: refactoring map expression.  generated code is broken right now

* wip

the current code shows a half-completed implementation of moving
generics out of ParselyWrite's type and into associated types.  Then
ParselyWrite can be used as a trait bound, which is useful when trying
to coerce types correctly when writing (e.g. the 'map' ambiguity
problem).  Doing this requires an additional type (WriteAdaptor) which
doesn't work in its current form: instead of a single impl defined in
parsely we'd need to generate impls dynamically for every type otherwise
we run into "can't-define-foreign-trait-on-foreign-type" problems when
we generate impls in a calling crate.

* wip: good progress on moving the generic values to the 'write' method
instead: it looks like this accomplishes enabling ParselyWrite as a
trait bound but without requiring the 'WriteAdaptor' type to still
enable a generic buffer type as opposed to a concrete one.  I still
think the 'Ctx' type might work best as an associated type and we don't
need generics there, so need to try that.

* i think this completes the core work to try and improve the ergonomics
around expressions that may or may not return a ParselyResult.

there's still cleanup to do (arguably the ParselyRead trait should be tweaked
to match ParselyWrite now)

* ci: update workflow file/fix some formatting issues