feat: trait rework (#9)
* refactor: change ParselyWrite trait

i found i had scenarios where i wanted to be able to use the
ParselyRead/ParselyWrite traits but not necessarily with
`BitBuf`/`BitBufMut`, so I wanted to be able to use them with different
buffer types.  In order to do that, I needed to move the buffer back to
a generic on the ParselyWrite trait.  Previously this was an issue with
how I was coercing the result of expressions in attributes: I needed
know generics there for it to work reasonably as a trait bound.  But I
found putting the buffer there should work alright: it makes sense to
require we can get a type that can be written to the current buffer.

* refactor: rename IntoParselyResult trait method: into_parsely_result_read -> into_parsely_result

* refactor: change ParselyRead trait definition to match new ParselyWrite

* feat: upgrade bits-io version, re-export BitSliceUxExts

* refactor: simplify alignment logic